fix: move all icons to svg files
