--- src/icons/AdobeIcon.js
@@ -1,13 +1,0 @@
-const AdobeIcon = ({ className }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 500 500"
-    className={className}
-  >
-    <path
-      d="M241.7 267.2h-26.3l-6.6 19.9c-.1.6-.5.9-1.2.9h-11.9c-.8 0-1-.4-.8-1.1l23.6-67.4c.5-1.1.9-3.3 1-6.4 0-.5.3-.7.8-.7h16.3c.4 0 .7.2.9.6l26.3 74c.2.7 0 1-.7 1h-13.4c-.5 0-.9-.3-1.1-.8l-6.9-20zm-23.4-11h20.4c-.3-1.1-1.9-5.9-4.8-14.4-2.8-8.6-4.7-14.5-5.6-17.7h-.1c-.8 3.9-4.1 14.6-9.9 32.1zM295.8 289.1c-8.8 0-15.8-2.4-21.2-7.1-5.3-4.8-7.9-11.9-7.9-21.2 0-8.4 2.8-15.3 8.5-20.8s13.4-8.3 23-8.3c2.6 0 4.8.2 6.4.7V207c0-.5.3-.8.8-.8h12.9c.5 0 .7.2.7.7v68.4c0 1.3.1 3.9.3 7.7.1.5-.2.9-.7 1-7.1 3.4-14.7 5.1-22.8 5.1zm8.7-11.8v-34.4c-1.7-.8-3.9-1.2-6.6-1.2-5 0-9 1.6-12 4.9-3 3.2-4.5 7.7-4.5 13.6 0 6.2 1.5 10.8 4.5 13.8s6.8 4.5 11.5 4.5c3.1.1 5.4-.3 7.1-1.2zM355.5 231.7c8.6 0 15.4 2.7 20.3 8.1 4.9 5.4 7.4 12.2 7.4 20.4 0 8.9-2.6 16-7.7 21.2-5.1 5.1-11.9 7.7-20.2 7.7-9.2 0-16.1-2.8-20.9-8.3-4.8-5.6-7.1-12.4-7.1-20.4 0-8.4 2.6-15.3 7.7-20.7 5.2-5.3 12-8 20.5-8zm-.6 9.9c-3.9 0-7.1 1.7-9.5 5-2.4 3.3-3.6 7.8-3.6 13.6s1.2 10.4 3.6 13.8c2.4 3.4 5.8 5.1 10.1 5.1 3.7 0 6.8-1.6 9.3-4.7 2.5-3.2 3.7-7.9 3.7-14.1 0-5.7-1.1-10.2-3.4-13.6-2.2-3.4-5.6-5.1-10.2-5.1zM392.2 206.2h12.7c.6 0 .9.3.9.9v26.7c3.7-1.4 7.6-2 11.8-2 7.9 0 14.2 2.5 18.9 7.6 4.8 5 7.1 11.5 7.1 19.4 0 9.5-2.9 16.9-8.8 22.4-5.9 5.4-12.9 8.1-21.1 8.1-8.5 0-15.7-1.4-21.7-4.3-.5-.2-.7-.6-.7-1.1V207c.1-.6.3-.8.9-.8zm36.7 53.9c0-5.8-1.3-10.2-4-13.2-2.6-3-6.1-4.5-10.6-4.5-3.3 0-6.2.6-8.6 1.8v33.1c2.2 1.2 5.1 1.8 8.7 1.8 4 0 7.4-1.6 10.2-4.9 2.9-3.3 4.3-8 4.3-14.1zM487 263h-22.8c.1 4.6 1.5 8.3 4.3 11.1 2.9 2.7 7 4.1 12.5 4.1 5.4 0 10.1-.9 14.1-2.8.5-.2.7 0 .7.6v8.5c0 .7-.2 1.1-.7 1.4-4.8 2.2-10.5 3.3-17.1 3.3-5.1 0-9.5-.8-13.3-2.4-3.8-1.6-6.7-3.7-8.8-6.5-2.1-2.7-3.7-5.7-4.8-8.9-1-3.3-1.5-6.7-1.5-10.4 0-8.2 2.4-15.2 7.1-20.8 4.8-5.7 11.4-8.5 19.7-8.5 7.9 0 13.8 2.3 17.7 7 3.9 4.7 5.9 10.2 5.9 16.5 0 2.3-.2 4.5-.5 6.5-.1.5-.4.8-.9.9-2.1.3-6 .4-11.6.4zm-22.9-8.8h16.1c3.6 0 5.6-.1 6-.2.1-.3.1-.8.1-1.6 0-2.9-.8-5.4-2.5-7.7-1.6-2.3-4.2-3.4-7.9-3.4-3.6 0-6.4 1.2-8.4 3.7-1.9 2.4-3 5.5-3.4 9.2zM94.4 183.8h55.3v132.5zM55.4 183.8H0v132.5zM74.9 232.6l35.2 83.7H87l-10.5-26.7H50.7z"
-    />
-  </svg>
-);
-
-export default AdobeIcon;
