--- src/icons/AirbnbIcon.js
@@ -1,11 +1,0 @@
-const AirbnbIcon = ({ className }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 500 500"
-    className={className}
-  >
-    <path d="M259.9 221.7c0 4.1-3.3 7.4-7.4 7.4-4.1 0-7.4-3.3-7.4-7.4 0-4.1 3.2-7.4 7.4-7.4 4.2.1 7.4 3.4 7.4 7.4zm-30.6 14.9v1.8s-3.5-4.6-11.1-4.6c-12.4 0-22.1 9.5-22.1 22.6 0 13 9.6 22.6 22.1 22.6 7.6 0 11.1-4.7 11.1-4.7v1.9c0 .9.7 1.6 1.6 1.6h9.2V235h-9.2c-.9 0-1.6.8-1.6 1.6zm0 27.5c-1.7 2.5-5.1 4.7-9.2 4.7-7.3 0-12.9-4.6-12.9-12.3 0-7.8 5.6-12.3 12.9-12.3 4 0 7.6 2.3 9.2 4.7v15.2zM247 235h10.9v42.9H247V235zm163.5-1.1c-7.5 0-11.1 4.6-11.1 4.6v-24.1h-10.9v63.5h9.2c.9 0 1.6-.8 1.6-1.6v-1.9s3.5 4.7 11.1 4.7c12.4 0 22.1-9.6 22.1-22.6 0-13.1-9.7-22.6-22-22.6zm-1.8 34.7c-4.2 0-7.5-2.2-9.2-4.7v-15.3c1.7-2.3 5.4-4.7 9.2-4.7 7.3 0 12.9 4.6 12.9 12.3 0 7.9-5.6 12.4-12.9 12.4zm-25.9-16.1V278h-10.9v-24.3c0-7.1-2.3-9.9-8.4-9.9-3.3 0-6.7 1.7-8.9 4.2v29.9h-10.9V235h8.7c.9 0 1.6.8 1.6 1.6v1.8c3.2-3.3 7.4-4.6 11.6-4.6 4.8 0 8.8 1.4 12 4.1 3.7 3.3 5.2 7.4 5.2 14.6zM317 233.9c-7.5 0-11.1 4.6-11.1 4.6v-24.1H295v63.5h9.2c.9 0 1.6-.8 1.6-1.6v-1.9s3.5 4.7 11.1 4.7c12.4 0 22.1-9.6 22.1-22.6.1-13.1-9.6-22.6-22-22.6zm-1.8 34.7c-4.2 0-7.5-2.2-9.2-4.7v-15.3c1.7-2.3 5.4-4.7 9.2-4.7 7.3 0 12.9 4.6 12.9 12.3 0 7.9-5.6 12.4-12.9 12.4zm-29.7-34.7c3.3 0 5 .6 5 .6v10.1s-9.1-3.1-14.8 3.4v30h-10.9v-43h9.2c.9 0 1.6.8 1.6 1.6v1.8c2.1-2.4 6.5-4.5 9.9-4.5zM171.8 274c-.6-1.4-1.1-2.9-1.7-4.1-.9-2.1-1.8-4-2.6-5.8l-.1-.1c-7.9-17.1-16.3-34.4-25.2-51.5l-.3-.7c-.9-1.7-1.8-3.5-2.7-5.4-1.1-2.1-2.3-4.2-4.1-6.3-3.6-4.6-8.9-7.1-14.5-7.1-5.7 0-10.8 2.5-14.6 6.8-1.7 2.1-3 4.2-4.1 6.3l-2.7 5.4-.3.7c-8.8 17.1-17.3 34.4-25.2 51.5l-.1.2c-.8 1.8-1.7 3.8-2.6 5.8-.6 1.3-1.1 2.6-1.7 4.1-1.5 4.2-1.9 8.2-1.4 12.3 1.3 8.6 7 15.7 14.8 18.9 3 1.3 6 1.8 9.2 1.8.9 0 2.1-.1 3-.2 3.8-.5 7.6-1.7 11.4-3.9 4.7-2.6 9.1-6.4 14.1-11.9 5 5.5 9.6 9.2 14.1 11.9 3.8 2.2 7.6 3.4 11.4 3.9.9.1 2.1.2 3 .2 3.2 0 6.4-.6 9.2-1.8 8-3.2 13.6-10.5 14.8-18.9.9-3.9.4-7.9-1.1-12.1zm-51.4 5.9c-6.2-7.8-10.1-15.1-11.5-21.2-.6-2.6-.7-4.9-.3-7 .2-1.8.9-3.4 1.8-4.8 2.2-3.1 5.8-5 10-5s8 1.8 10 5c.9 1.4 1.6 3 1.8 4.8.3 2.1.2 4.4-.3 7-1.3 6.1-5.3 13.4-11.5 21.2zm45.5 5.4c-.8 5.9-4.8 11.1-10.4 13.3-2.7 1.1-5.7 1.5-8.7 1.1-2.9-.3-5.7-1.3-8.7-3-4.1-2.3-8.2-5.8-13-11.1 7.5-9.2 12.1-17.7 13.8-25.2.8-3.5.9-6.7.6-9.7-.5-2.9-1.5-5.5-3.1-7.8-3.5-5.1-9.5-8.1-16.1-8.1-6.6 0-12.5 3.1-16.1 8.1-1.6 2.3-2.6 4.9-3.1 7.8-.5 3-.3 6.3.6 9.7 1.7 7.5 6.4 16.1 13.8 25.3-4.7 5.2-8.9 8.8-13 11.1-3 1.7-5.8 2.6-8.7 3-3.1.3-6-.1-8.7-1.1-5.6-2.3-9.6-7.4-10.4-13.3-.3-2.9-.1-5.7 1-8.9.3-1.1.9-2.3 1.5-3.6.8-1.8 1.7-3.8 2.6-5.7l.1-.2c7.9-17 16.3-34.3 25.1-51.2l.3-.7c.9-1.7 1.8-3.5 2.7-5.2.9-1.8 1.9-3.5 3.2-5 2.4-2.7 5.6-4.2 9.1-4.2s6.7 1.5 9.1 4.2c1.3 1.5 2.3 3.2 3.2 5 .9 1.7 1.8 3.5 2.7 5.2l.3.7c8.7 17 17.1 34.3 25 51.3v.1c.9 1.8 1.7 3.9 2.6 5.7.6 1.4 1.1 2.5 1.5 3.6 1.3 3 1.7 5.8 1.2 8.8z" />
-  </svg>
-);
-
-export default AirbnbIcon;
