--- src/icons/ChevronLeftIcon.js
@@ -1,11 +1,0 @@
-const ChevronLeftIcon = ({ className }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 404.258 404.258"
-    className={className}
-  >
-    <path d="M289.927 18l-24-18-151.596 202.129 151.596 202.129 24-18-138.096-184.129z" />
-  </svg>
-);
-
-export default ChevronLeftIcon;
