--- src/icons/CiscoIcon.js
@@ -1,11 +1,0 @@
-const CiscoIcon = ({ className }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 512 512"
-    className={className}
-  >
-    <path d="M348.4 387.7V364l-.5.3c-5.3 2.6-11.2 4.1-17.1 4.3-6.2.5-12.3-1.3-17.2-5.1-3.5-2.7-6.2-6.4-7.5-10.6-1.5-4.7-1.7-9.6-.5-14.3 1.1-4.8 3.7-9.1 7.6-12.2 3.4-2.8 7.6-4.6 12-5.2 7.8-.8 15.7.7 22.7 4.3l.5.3V302l-1.1-.3c-7-2.3-14.3-3.2-21.6-2.8-6.7.2-13.3 1.7-19.3 4.4-6.2 2.8-11.6 6.9-15.7 12.2-13.3 16.8-13.7 40.5-.9 57.7 6.3 8.1 15.2 13.8 25.1 16.2 10.5 2.6 21.5 2.2 31.7-1.1l1.8-.6zm-234.6 0V364l-.5.3c-5.3 2.6-11.1 4.1-17.1 4.3-6.2.5-12.3-1.3-17.2-5.1-3.5-2.7-6.1-6.4-7.5-10.6-1.5-4.6-1.7-9.6-.5-14.3 1.1-4.8 3.7-9.1 7.6-12.2 3.4-2.8 7.6-4.6 12-5.2 7.8-.8 15.7.7 22.7 4.3l.5.3V302l-1.1-.3c-7-2.3-14.3-3.2-21.6-2.8-6.7.2-13.2 1.7-19.3 4.4-6.1 2.8-11.5 6.9-15.7 12.2-13.3 16.8-13.7 40.5-.9 57.7 6.3 8.1 15.2 13.8 25.1 16.2 10.5 2.6 21.4 2.2 31.7-1.1l1.8-.6zm83.8-19.7l-.1 20.4 1.4.3c4.8.8 9.4 1.4 14.2 1.8 4.2.3 8.4.4 12.6.3 4-.3 7.9-1.1 11.7-2.3 4.5-1.3 8.6-3.5 12.2-6.5 3.2-2.8 5.7-6.3 7.1-10.2 2.6-6.7 2.5-13.9-.2-20.5-1.3-3.4-3.5-6.5-6.3-8.8-2.7-2.4-5.9-4.3-9.3-5.4l-15.4-5.7c-1.8-.8-3.3-2.3-4.2-4-.7-1.5-.7-3.3.2-4.7l3.3-3.1c1.7-.8 3.6-1.3 5.5-1.6 3.3-.4 6.6-.4 9.8.2 3.8.4 7.6 1.1 11.3 2l.8.2v-18.8l-1.5-.3c-3.8-.8-7.7-1.6-11.6-2-5-.7-10.2-.8-15.3-.1-6.3.6-12.4 3-17.5 7-3.4 2.8-6 6.3-7.5 10.3-2.2 6.2-2.2 13 0 19.2 1.2 3.2 3 5.9 5.4 8.2 2.1 2 4.4 3.7 7 5 3.3 1.8 6.9 3.2 10.6 4.2l3.6 1.2 1.8.7.5.3c1.8.6 3.5 1.7 4.8 2.9l2.2 3.3v3.4c-.6 1.8-1.8 3.3-3.5 4.3-1.8.9-3.8 1.6-5.9 1.8-3 .5-6 .6-9.1.4-4-.3-7.9-.8-11.9-1.6-2-.4-4.1-.9-6.1-1.4l-.6-.4zm-52.8 21.2h22.4v-88.8h-22.4v88.8zm259.4-62.6c10.1-8.2 24.8-6.7 32.9 3.3 8.2 10 6.7 24.8-3.3 32.9l-.4.3c-24.3 18.7-52.7-17-29.2-36.5zm-32.5 18.2c.5 29.7 28 51.4 56.9 45 41.2-7.8 50.9-62.4 14.9-84-31-19.2-71.1 2.5-71.8 39zm118.1-116c.6 14.2 21.6 14.2 22.2 0v-24.6c-.6-14.2-21.6-14.2-22.2 0v24.6zm-60.6 0c1 13.7 21.2 13.7 22.2 0v-55.2c-.8-13.9-21.3-13.9-22.2 0v55.2zm-61.6 22.3c1.1 13.7 21.2 13.7 22.2 0V131.7c-.6-14.2-21.5-14.2-22.2 0v119.4zm-60.5-22.3c.8 13.9 21.3 13.9 22.2 0v-55.2c-.7-14.2-21.6-14.2-22.2 0v55.2zm-61.9 0c1.3 13.4 20.9 13.4 22.2 0v-24.6c-.6-14.2-21.5-14.2-22.2 0v24.6zm-61.2 0c.6 14.2 21.5 14.2 22.2 0v-55.2c-.7-14.2-21.6-14.2-22.2 0v55.2zm-61.6 22.3c.6 14.2 21.5 14.2 22.2 0V131.7c-.7-14.2-21.6-14.2-22.2 0v119.4zm-61.1-22.3c1.3 13.6 21 13.6 22.2 0v-55.2c-.6-14.2-21.5-14.2-22.1 0l-.1 55.2zm-61.3 0c1 13.7 21.1 13.7 22.2 0v-24.6C21.6 190 .6 190 0 204.2v24.6z" />
-  </svg>
-);
-
-export default CiscoIcon;
