--- src/icons/CloudIcon.js
@@ -1,11 +1,0 @@
-const CloudIcon = ({ className }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 512 512"
-    className={className}
-  >
-    <path d="M387.52 197.56C372.69 127.51 306.55 76 227 76 125.74 76 51.58 157.94 60.64 236.19 24.48 253.18 0 290.48 0 331c0 57.89 47.11 105 105 105h287c66.34 0 120-53.67 120-120 0-64.74-51.31-121.01-124.48-118.44z" />
-  </svg>
-);
-
-export default CloudIcon;
