--- src/icons/EbayIcon.js
@@ -1,11 +1,0 @@
-const EbayIcon = ({ className }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 512 512"
-    className={className}
-  >
-    <path d="M128.6 153.6v99.7c-3-19.7-15.5-48.5-62.4-48.5-64 0-66.5 49.1-66.1 61.9 0 0-3.2 57.6 64.5 57.6 50.1 0 59.2-26.7 60.8-34.1H98.2c-2.1 4.9-9.9 18.1-33.6 17.6-32.3-.6-37.5-27.5-38.4-34.7h102.4v51.2h25.1v-15.5s12.7 15.5 46.9 15.5c29 0 52.7-15.3 59.7-43.7 1.3-5.1 2-10.1 2.1-16 .7-38.9-29.1-59.5-61.3-59.7-32.2-.2-47.5 21.9-47.5 21.9v-73.1h-25zm131.7 126.9c-5.3 33.6 25.2 43.9 49.1 43.7s38.9-7.1 49.1-19.7v19.7h26.1v-74.1c-.7-12.6-2.5-43.1-58.1-45.3 0 0-53.1-4.8-58.7 34.1H295s2.2-17.6 29.9-17.1c26.1.5 33.7 18.6 33.6 34.1 0 0-25.6-.1-33.1 0-13.5.2-58.2-1.5-65.1 24.6zm107.2-75.7l59.2 110.4-22.9 43.2h28.8L512 204.8h-27.2l-43.7 85.3-43.7-85.3h-29.9zM65.6 221.9c36.7 0 36.8 34.1 36.8 34.1H26.2s2.6-34.1 39.4-34.1zm129.6 0c41.4-1.1 40 42.7 40 42.7s.7 42.1-40 42.7c-40.7.5-40-43.2-40-43.2s-1.4-41.2 40-42.2zM320 273.1c3.7-.2 6.4 0 6.4 0h29.3s4.8 35.8-41.6 35.2c0 0-26.2.1-27.7-20.3 0-11.3 22.4-14.3 33.6-14.9z" />
-  </svg>
-);
-
-export default EbayIcon;
