--- src/icons/IntelIcon.js
@@ -1,11 +1,0 @@
-const IntelIcon = ({ className }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 512 512"
-    className={className}
-  >
-    <path d="M150.8 230.4v82.7h-26.3V209.2h54.3c23.1 0 30.9 16.3 30.9 31.1v72.8h-26.3v-72.7c0-6.3-3.2-10-10.9-10h-21.7zm179-1.2c-9.1 0-16 4.6-18.9 10.9-1.7 3.5-2.5 7.3-2.5 11.2h40.5c-.5-11.2-5.6-22.1-19.1-22.1zm-21.4 40.1c0 13.4 8.3 23.4 23.3 23.4 11.6 0 17.3-3.2 24.1-10l16.3 15.6c-10.3 10.2-21.3 16.5-40.6 16.5-25.2 0-49.2-13.7-49.2-53.7 0-34.3 21-53.6 48.7-53.6 28.1 0 44.3 22.7 44.3 52.3v9.4l-66.9.1zM260 312.9c-21.4 0-30.5-14.9-30.5-29.7V180.8h26.3v28.3h19.7v21.2h-19.7v51.2c0 6.1 2.8 9.4 9.1 9.4h10.7v21.9l-15.6.1zM101.1 194.8H74.6v-25.2h26.5v25.2zm.1 119.4c-19.8-1.9-26.6-13.9-26.6-27.8v-77.2h26.5l.1 105zm315.8-2.3c-19.8-1.9-26.5-13.9-26.5-27.7V166.1H417v145.8zm92.7-128.7C485.7 66 258.1 58.5 111.5 147.8v9.9c146.4-75.3 354.3-74.8 373.2 33 6.2 35.7-13.8 72.9-49.7 94.3v28c43.2-15.8 87.6-67.2 74.7-129.8zm-266.5 205c-101.2 9.3-206.7-5.4-221.4-84.4-7.2-39 10.7-80.3 34.1-106v-13.7C13.5 221.3-9.5 268.3 3.7 323.9c16.9 71.3 107.4 111.7 245.5 98.3 54.7-5.3 126.3-22.9 175.9-50.3v-38.7c-45 27-119.8 49.2-181.9 55zm204.5-214.4c0-1.2-.8-1.7-2.5-1.7h-1.7v3.7h1.7c1.7 0 2.5-.5 2.5-1.7v-.3zm2.5 9.1h-2l-.4-.3-2.7-4.6-.4-.2h-1.2v4.5l-.5.5h-1.7l-.5-.5V171l.8-1 3.2-.1c3.2 0 5.2.9 5.2 4v.2c0 1.9-1 2.9-2.4 3.4l2.7 4.7.1.3-.2.4zm-4.6-15.2c-4.8 0-8.7 3.9-8.7 8.7.1 4.8 4 8.8 8.8 8.8 4.8 0 8.7-3.9 8.7-8.7-.1-4.8-4-8.8-8.8-8.8zm0 19.4c-5.8 0-10.6-4.7-10.6-10.5s4.8-10.6 10.6-10.6c5.7 0 10.6 4.7 10.6 10.6 0 5.8-4.9 10.5-10.6 10.5z" />
-  </svg>
-);
-
-export default IntelIcon;
