--- src/icons/MouseClickIcon.js
@@ -1,14 +1,0 @@
-const MouseClickIcon = ({ className }) => (
-  <svg
-    viewBox="0 0 24 24"
-    xmlns="http://www.w3.org/2000/svg"
-    className={className}
-  >
-    <path d="M.5 4a.5.5 0 01-.5-.5v-1C0 1.122 1.122 0 2.5 0h1a.5.5 0 010 1h-1C1.673 1 1 1.673 1 2.5v1a.5.5 0 01-.5.5zM19.5 4a.5.5 0 01-.5-.5v-1c0-.827-.673-1.5-1.5-1.5h-1a.5.5 0 010-1h1C18.878 0 20 1.122 20 2.5v1a.5.5 0 01-.5.5zM12.5 1h-5a.5.5 0 010-1h5a.5.5 0 010 1zM9.5 20h-2a.5.5 0 010-1h2a.5.5 0 010 1zM3.5 20h-1A2.503 2.503 0 010 17.5v-1a.5.5 0 011 0v1c0 .827.673 1.5 1.5 1.5h1a.5.5 0 010 1zM.5 13a.5.5 0 01-.5-.5v-5a.5.5 0 011 0v5a.5.5 0 01-.5.5zM19.5 10a.5.5 0 01-.5-.5v-2a.5.5 0 011 0v2a.5.5 0 01-.5.5z" />
-    <g>
-      <path d="M19.593 24a.5.5 0 01-.459-.304l-1.901-4.442-2.267 1.975a.624.624 0 01-.433.162.663.663 0 01-.663-.663V8.663c0-.52.665-.852 1.094-.506l8.844 8.84c.138.167.192.315.192.47a.663.663 0 01-.663.663h-3.012l1.828 4.273a.499.499 0 01-.263.656l-2.1.9a.497.497 0 01-.197.041zm-2.171-6.074a.499.499 0 01.459.304l1.975 4.614 1.181-.506-1.929-4.51a.501.501 0 01.459-.697h2.964L14.87 9.473v10.515l2.224-1.938a.494.494 0 01.328-.124zm-3.097 2.536a.339.339 0 00-.045.04zm8.74-2.798h.01z" />
-    </g>
-  </svg>
-);
-
-export default MouseClickIcon;
