--- src/icons/NikeIcon.js
@@ -1,11 +1,0 @@
-const TargetIcon = ({ className }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 500 500"
-    className={className}
-  >
-    <path d="M101.9 317.7c-11.3-.4-20.6-3.5-27.8-9.3-1.4-1.1-4.7-4.4-5.8-5.8-2.9-3.7-4.9-7.3-6.3-11.3-4.1-12.3-2-28.4 6-46.1 6.9-15.1 17.5-30.1 36-50.9 2.7-3.1 10.8-12 10.9-12 0 0-.4.8-1 1.7-4.8 8.1-8.9 17.5-11.2 25.8-3.6 13.2-3.2 24.5 1.3 33.3 3.1 6 8.3 11.3 14.2 14.2 10.3 5.1 25.5 5.5 43.9 1.2 1.3-.3 64.3-17 140-37.2C378 201.1 440 184.6 440 184.6S264 259.9 172.6 299c-14.5 6.2-18.3 7.7-25.1 10.1-17.4 6.1-33 9.1-45.6 8.6z" />
-  </svg>
-);
-
-export default TargetIcon;
