--- src/icons/PixarIcon.js
@@ -1,11 +1,0 @@
-const PixarIcon = ({ className }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 500 500"
-    className={className}
-  >
-    <path d="M122.6 253.7c0-.8.1-.8 2.7-1.1l2.5-.3.1-.6c.1-.7.1-14.7 0-23.1l-.1-5.9-.7-.2c-.4-.1-1.5-.3-2.5-.3-1.8-.1-2-.2-2.1-.8 0-.2.4-.3 6.9-.3 10.3 0 13.3.3 16.1 1.8 2.1 1.1 3.5 2.9 3.7 4.8.2 1.7-.6 4.3-1.8 6-1.7 2.5-5.3 4.8-9 5.7-2.1.6-2.5.6-2.5.3 0-.8.2-.9 1.6-1.5 1.8-.7 3-1.5 3.9-2.5 1.5-1.7 2.1-3.6 2-6.3-.1-1.6-.1-1.8-.5-2.8-1.2-2.5-3.7-3.9-7.7-4.1l-1.6-.1-.1.7c-.1.7-.1 24 0 27.1.1 1.9.1 2 .4 2.1.2 0 1.5.2 2.9.3 1.4.1 2.7.3 2.8.3.1 0 .2.2.2.4v.4H133c-3.7 0-7.6 0-8.6.1-1.5.1-1.8.1-1.8-.1zM178.6 253.5c0-.2.1-.4.1-.5.1-.1.9-.2 1.9-.3 1-.1 2.1-.3 2.5-.4.7-.2.7-.2.8-.9.2-1.6.1-28-.2-28.4-.1-.1-1.2-.3-2.5-.5s-2.4-.3-2.4-.3-.1-.2-.1-.4l-.1-.4h7c3.8 0 7.5 0 8.1-.1l1.2-.1-.1.4c-.1.2-.2.5-.2.5-.1.1-1.2.2-2.5.4s-2.4.3-2.5.4c-.2.2-.2 20.9-.1 25.5l.1 3.7.7.2c.4.1 1.5.3 2.5.3.9.1 1.7.2 1.8.2s.1.2.1.4l.1.4H187c-4.3 0-7.9.1-8.1.1-.2.1-.3.1-.3-.2zM253.9 254.2c-2.6-2.1-7.2-7.1-13.1-14.1-.6-.8-1.2-1.4-1.3-1.4-.5 0-6.6 9.2-8.2 12.6l-.5.9 1.5.2c.8.1 1.7.3 1.9.3.4.1.7.3.7.7 0 .1-1.6.2-6.1.2h-6.1l.1-.3c.2-.5.4-.6 1.6-.8 1.6-.2 2-.5 3.3-1.9 1.6-1.7 9-11.1 10.6-13.4.1-.1-.1-.5-.9-1.5-.6-.7-2.6-3.4-4.6-6-1.9-2.6-4-5.3-4.6-5.9-1-1.1-1.1-1.2-2-1.5-.5-.1-1.2-.3-1.5-.3-.5 0-.6-.1-.6-.4v-.4s3.3-.1 7.3-.1c6.8-.1 7.3-.1 7.3.2 0 .5-.2.6-1.8.8-.9.1-1.8.3-2 .4l-.4.1 1 1.6c2 3.1 6.2 8.8 6.4 8.9.3 0 5.6-7.5 6.9-9.7.5-.8.4-.8-1.5-1.2-1.4-.2-1.5-.3-1.5-.6 0-.2-.1-.4-.1-.4s2.5 0 5.5-.1l5.5-.1-.1.5c-.1.2-.1.5-.1.5s-.7.2-1.5.3c-1.4.3-1.5.3-2.3 1.1-.9.9-4.9 5.4-7.8 8.8l-2 2.3 1.2 1.5c1.2 1.5 1.7 2.2 7.1 9.9 2.3 3.3 5.8 8 6.4 8.8.1.1-.2.2-1.3.2h-1.4l-1-.7zM374.8 253.9c-2.5-1.9-7.9-7.2-13.2-13.2l-2.1-2.4-2.2.1c-1.2 0-2.2.1-2.3.2-.1.1 0 12.9.2 13.2.1.2 1 .4 3.1.6 1.7.2 1.8.2 1.9.6.1.2.1.4 0 .4 0 0-3.6.1-8 .1l-8 .1c0-.3 0-.5.2-.8.1-.1 1.2-.3 2.5-.4 1.3-.1 2.4-.3 2.4-.4.1-.1.2-.7.2-1.4.1-1.4-.1-26.7-.2-27.4-.1-.5-.6-.7-2.5-.8-2.5-.2-2.6-.2-2.6-.7v-.4h5.4c15.1 0 18.4.5 20.8 3.1 1.5 1.6 1.6 3.8.4 6.2-.9 1.8-3.2 4.2-5.1 5.3l-.8.4.5.9c.8 1.3 5.1 7.7 6.8 10.2 1.8 2.5 2.7 3.7 4.4 5.6l1.2 1.4-.9.2-1 .2c.1 0-.5-.4-1.1-.9zM359 236.7c2.1-.4 3.5-1.1 4.8-2.4 1.2-1.2 1.8-2.2 2.1-3.9.7-3.5-1.6-6.3-5.9-7.4-1.6-.4-4.6-.7-4.9-.5-.1.1-.2 2-.2 7.4v7.3l1.3-.1c.7-.1 1.9-.3 2.8-.4zM280.1 253.3c.1-.5-.1-.5 2.4-.9l1.7-.3.6-1.1c1.1-2 3.3-7.5 7.9-19.4 1.9-4.8 3.4-9 3.3-9 0 0-.7-.2-1.5-.3-2.3-.4-2.2-.3-2.2-.8v-.4h7.2c6.6 0 7.1 0 7.1.2-.1.1-.1.3-.1.4 0 .1-.6.3-1.7.5-1 .1-1.8.3-1.8.4 0 .1.7 2.4 1.7 5.2 3.6 10.2 6.6 18 8.7 22.4l.9 1.8.7.2c.7.2 2.7.5 3.2.5.2 0 .3.2.3.4l.1.4-7.9.1c-4.3 0-7.9.1-7.9.1s0-.2.1-.5l.1-.5 2.1-.3c2.6-.3 2.8-.4 2.6-.7-.2-.4-1.5-4.1-2.4-6.8-.5-1.4-1.2-3.4-1.6-4.5l-.8-2h-.9c-2.3 0-10 .5-10.2.6-.3.3-2.4 7.9-3 10.8-.2 1.1-.4 2-.4 2 0 .1.7.3 2.4.5 1.7.3 1.7.3 1.8.7v.4h-5.5c-3 0-5.9 0-6.3.1l-.8.1.1-.3zm22.2-16.5c-.1-.2-.2-.6-.4-.8-.2-.4-2.1-6-3.6-10.5-.4-1.1-.7-2-.7-2.1-.2-.2-5 12.8-5 13.5 0 .1 1.7.1 3.8.2 2.1 0 4.3.1 5 .1h1.2l-.3-.4zM67.5 277.5c0-.1.1-.1.5-.2.7-.1.6-.1.8-.5.3-.6.7-1.7 1.3-3.2.6-1.6 1.8-4.8 1.8-4.9 0 0-.1-.1-.5-.1-.3 0-.5-.1-.5-.1v-.2h3.6v.1c0 .1 0 .1-.5.2-.3 0-.5.1-.5.1s.1.4.3.9c.8 2.6 2.1 5.8 2.7 7.3l.2.3h.2c.1 0 .4.1.6.1l.4.1v.1h-3.8c0-.1.1-.2.3-.2.4 0 .9-.1 1-.2 0 0 0-.3-.3-.9-.4-1.1-.8-2.4-1-2.8l-.1-.3h-.8c-.8 0-1.7.1-2.2.1h-.2l-.2.5c-.5 1.6-.9 3.2-.8 3.3 0 0 .7.2.9.2.1 0 .1 0 .1.1v.1h-1.6l-1.7.1zm6.3-4.5c0-.1-.9-2.6-1.2-3.5l-.2-.6-.1.1c-.1.2-.4 1-.8 2.1-.5 1.4-.6 1.7-.6 1.8h2.5c.2.1.4.1.4.1zM98.4 276c-1.9-2.4-5-6.2-5.2-6.4l-.1-.1v1.8c0 3.7.1 5.9.2 5.9 0 0 .4.1 1 .1h.4v.2h-3.4v-.1s.1-.1.4-.1c.5-.1.7-.1.7-.2.1-.2.2-2.8.3-6.5v-1.5s-.2 0-.9-.1c-.3 0-.5 0-.5-.1v-.2h2.5l.2.3c.6.9 2.1 2.8 4.1 5.3.7.9 1.4 1.8 1.5 1.9l.1.2v-.7c0-1.6-.1-6-.1-6.4 0-.1 0-.1-.2-.2-.1 0-.4-.1-.7-.1-.3 0-.5-.1-.5-.1v-.2h1.7c1.7 0 1.7 0 1.7.1v.1s-.2.1-.5.1-.5.1-.5.1-.1.8-.1 1.4c0 .7-.1 2.4-.1 5.3v2.4c-.2.2-.2.1-2-2.2zM256.4 276c-1.9-2.4-5.1-6.3-5.3-6.5l-.1-.1v1.8c0 3.7.1 6 .2 6 0 0 .4.1 1 .1h.4v.2H249.2v-.1s.1-.1.4-.1c.5-.1.7-.1.7-.2.1-.2.2-2.9.3-6.6V269s-.2-.1-.9-.1c-.3 0-.5-.1-.5-.1v-.2h2.6l.2.3c.6.9 2.2 2.8 4.2 5.4.8.9 1.4 1.8 1.5 1.9l.1.2v-.8c0-1.7-.1-6.1-.1-6.5 0-.1 0-.1-.2-.2-.1 0-.4-.1-.7-.1-.3 0-.5-.1-.5-.1v-.2h1.7c1.7 0 1.7 0 1.7.1v.1s-.2.1-.5.1-.5.1-.5.1-.1.8-.1 1.5-.1 2.4-.1 5.4v2.4c-.2.2-.3.1-2.1-2.2zM116.2 277.6c.1-.2.1-.2.2-.2.2 0 1.2-.1 1.2-.2.1-.1.1-8.1 0-8.2 0 0-.3-.1-1.1-.2-.2 0-.2 0-.2-.1v-.1h1.9c1.9 0 1.9 0 1.9.1v.1s-.3.1-.7.1l-.6.1v7.9c0 .3 0 .3.8.4.6.1.6.1.6.2h-1.9c-1 0-2.1.1-2.1.1zM134.9 277.6c0-.1.1-.1.4-.2.6-.1.6-.1.7-.2.2-.6.6-7.9.6-8.1 0 0-.2-.1-.7-.1-.2 0-.4-.1-.4-.1v-.2h3.2v.1s-.2.1-.4.1-.4.1-.4.1 0 .1.1.5c.5 1.4 2.7 6.1 2.7 6.1.1-.1 3.2-6.6 3.2-6.6s-.3-.1-.6-.1c-.3-.1-.4-.1-.4-.2h3v.1s-.1.1-.4.1c-.2 0-.4.1-.5.1h-.1v.3c.2 1.1.2 7.3.3 7.9 0 .1.1.2.6.2.3.1.4.1.4.2h-3.6v-.1c0-.1 0-.1.3-.1.4 0 1.1-.1 1.1-.2v-2.3V269.7l-.1.2c-.2.3-1 1.8-2.1 4-1 2.2-1.5 3.3-1.5 3.4l-.1.1v-.1c-.1-.1-1.1-3-2.3-5.7-.5-1.2-1.1-2.1-1.1-2.1l.1 1.7c0 1.8-.1 2.9-.1 4.1 0 .8-.1 1.8-.1 1.8s.3.1.7.1c.7.1.7.1.7.2v.1h-1.6c-1.4.2-1.7.2-1.6.1zM291.1 278.7c.1-.8.2-1.7.2-2.9v-1.1h.2v.4c0 .7.1.8.2 1.1.2.3.6.6.9.8.4.2.7.3 1.1.3.4 0 .7-.1 1-.2.7-.4 1.1-1.2.9-2-.2-.6-.5-.9-1.6-1.7-1-.7-1.3-.9-1.7-1.3-.7-.7-.8-1.6-.4-2.4.2-.5.5-.8 1-1 .4-.2.7-.3 1.3-.3.6 0 1.1.1 1.4.2.1 0 .2.1.3.1 0 0 .1-.1.2-.5l.1-.5h.2s0 .2-.1.5c-.1.8-.2 1.7-.2 2.3 0 .5 0 .5-.1.5-.2-.1-.2 0-.2-.5 0-.6-.1-.8-.4-1.1-.3-.3-.5-.4-.9-.5-.3-.1-.9 0-1.1 0-.7.2-1.1.7-1 1.4 0 .6.3.9 1.8 1.9.8.5 1.1.8 1.4 1.1.4.4.6.8.7 1.3.1.3 0 1-.1 1.2-.3 1-1.1 1.7-2.3 1.9-.6.1-1.5 0-2.2-.2-.1 0-.3-.1-.3-.1s-.1.3-.1.6c-.2.7-.1.7-.2.7 0 .2 0 .2 0 0zM228.9 277.8c-1.2-.1-2.1-.6-2.9-1.3-.7-.7-1-1.4-1.2-2.2-.1-.5-.1-1.4 0-1.8.2-1 .6-1.7 1.3-2.4 1.2-1.2 2.7-1.7 4.6-1.5.8.1 1.4.2 2 .5 1.2.6 2 1.6 2.3 2.9.1.3.1 1.1 0 1.5-.3 2.4-2.1 4.1-4.6 4.4-.3 0-1 0-1.5-.1zm2.2-.5c1.6-.3 2.7-1.3 3-3 .1-.5.1-1.3 0-1.7-.4-1.8-1.8-3.2-3.6-3.6-.4-.1-1.4-.1-1.8 0-1.9.4-3.1 1.8-3.1 3.8 0 1.6 1 3.3 2.6 4.1 1 .4 1.9.6 2.9.4zM358.7 277.7c.1-.2.1-.2.4-.2.4 0 .9-.1 1-.2.1-.1.1-8 0-8.2 0 0-.6-.1-1.1-.2-.2 0-.3-.1-.3-.1h2.9c3 0 3.1 0 4 .2 1.1.2 1.8.5 2.5 1.1.3.3.5.6.7.9.3.5.3 1.1.3 1.8-.2 1.8-1.2 3.2-2.9 4.1-.9.4-1.9.7-3.1.8h-4.4zm4.9-.5c1.3-.1 2.5-.7 3.3-1.5.5-.5.9-1.1 1-1.8.1-.3.1-.4.1-1s0-.7-.1-1c-.1-.4-.4-.9-.6-1.2-.9-1.2-2.4-1.7-4.9-1.7h-1.3v.1c0 .1 0 1-.1 2.1v3.7c0 1 0 1.8.1 1.9 0 .1.1.2.6.2.7.3 1.3.3 1.9.2zM185 277.4c0-.1.4-.1.8-.2.4 0 .8-.1.8-.1v-.9-7h-.7c-1 0-1.7.1-2.3.2-.3.1-.4.1-.5.2 0 0-.1.2-.1.5-.1.7-.1.7-.2.7h-.1s0-.3.1-.6c.1-.8.1-1.7.1-2v-.3h.2V268.7h.3c1.7.1 6.5.1 7.7 0h.2l.1-.3c0-.2.1-.3.1-.4 0-.1 0-.1.1-.1h.2c0 .1-.1 1-.3 2-.1.5-.1 1-.1 1h-.2v-1.4l-.2-.1c-.5-.2-1.6-.3-2.7-.3h-.6v7.9h.1s.4 0 .8.1c.4 0 .7.1.7.1v.2H185zM338.8 277.5c-1.3-.2-2.2-.9-2.6-2.1-.2-.6-.2-.6-.2-3.4v-2.7c0-.1 0-.1-.6-.2-.3-.1-.4-.1-.4-.2h3.4v.1s-.3.1-.7.1c-.3 0-.6.1-.6.1s-.1 2.3-.1 3.5c0 1.4 0 1.7.2 2.3.3 1.3 1.2 2.1 2.3 2.1.7 0 1.3-.1 1.7-.3.6-.3 1-.9 1.2-1.6.1-.4.1-.6.1-1.5 0-.8 0-4.5-.1-4.5 0 0-.8-.1-1.1-.2-.2 0-.2 0-.3-.1v-.1h3.2v.1c0 .1-.1.1-.4.2-.2 0-.4.1-.4.1 0 .1-.1 1.7-.1 3 0 1.4 0 1.9-.1 2.5-.3 1.6-1.4 2.6-3 2.8h-1.4zM160.5 277.5c0-.1.1-.1.5-.2.7-.1.6-.1.8-.5.3-.6.7-1.7 1.3-3.2.6-1.6 1.8-4.8 1.8-4.9 0 0-.1-.1-.5-.1-.3 0-.5-.1-.5-.1v-.2h3.6v.1c0 .1 0 .1-.5.2-.3 0-.5.1-.5.1s.1.4.3.9c.8 2.6 2.1 5.8 2.7 7.3l.2.3h.2c.1 0 .4.1.6.1l.4.1v.1h-3.8c0-.1.1-.2.3-.2.4 0 .9-.1 1-.2 0 0 0-.3-.3-.9-.4-1.1-.8-2.4-1-2.8l-.1-.3h-.8c-.8 0-1.7.1-2.2.1h-.2l-.2.5c-.5 1.6-.9 3.2-.8 3.3 0 0 .7.2.9.2.1 0 .1 0 .1.1v.1h-1.6l-1.7.1zm6.2-4.5c0-.1-.9-2.6-1.2-3.5l-.2-.6-.1.1c-.1.2-.4 1-.8 2.1-.5 1.4-.6 1.7-.6 1.8h2.5c.3.1.4.1.4.1zM206.3 277.6c.1-.2.1-.2.2-.2.2 0 1.2-.1 1.2-.2.1-.1.1-8.1 0-8.2 0 0-.3-.1-1.1-.2-.2 0-.2 0-.2-.1v-.1h1.9c1.9 0 1.9 0 1.9.1v.1s-.3.1-.7.1l-.6.1v7.9c0 .3 0 .3.8.4.6.1.6.1.6.2h-1.9c-1.2.1-2.1.1-2.1.1zM313.9 277.4c0-.1.4-.1.8-.2.4 0 .8-.1.8-.1v-.9-7h-.7c-1 0-1.7.1-2.3.2-.3.1-.4.1-.5.2 0 0-.1.2-.1.5-.1.7-.1.7-.2.7h-.1s0-.3.1-.6c.1-.8.1-1.7.1-2v-.3h.2V268.7h.3c1.7.1 6.5.1 7.7 0h.2l.1-.3c0-.2.1-.3.1-.4 0-.1 0-.1.1-.1h.2c0 .1-.1 1-.3 2-.1.5-.1 1-.1 1h-.2v-1.4l-.2-.1c-.5-.2-1.6-.3-2.7-.3h-.6v7.9h.1s.4 0 .8.1c.4 0 .7.1.7.1v.2H313.9zM406.2 277.8c-1.2-.1-2.1-.6-2.9-1.3-.7-.7-1-1.4-1.2-2.2-.1-.5-.1-1.4 0-1.8.2-1 .6-1.7 1.3-2.4 1.2-1.2 2.7-1.7 4.6-1.5.8.1 1.4.2 2 .5 1.2.6 2 1.6 2.3 2.9.1.3.1 1.1 0 1.5-.3 2.4-2.1 4.1-4.6 4.4-.3 0-1.1 0-1.5-.1zm2.2-.5c1.6-.3 2.7-1.3 3-3 .1-.5.1-1.3 0-1.7-.4-1.8-1.8-3.2-3.6-3.6-.4-.1-1.4-.1-1.8 0-1.9.4-3.1 1.8-3.1 3.8 0 1.6 1 3.3 2.6 4.1 1 .4 1.9.6 2.9.4zM383.6 277.6c.1-.2.1-.2.2-.2.2 0 1.2-.1 1.2-.2.1-.1.1-8.1 0-8.2 0 0-.3-.1-1.1-.2-.2 0-.2 0-.2-.1v-.1h1.9c1.9 0 1.9 0 1.9.1v.1s-.3.1-.7.1l-.6.1v7.9c0 .3 0 .3.8.4.6.1.6.1.6.2h-1.9c-1 0-2.1.1-2.1.1zM427.1 278.7c.1-.8.2-1.7.2-2.9v-1.1h.2v.4c0 .7.1.8.2 1.1.2.3.6.6.9.8.4.2.7.3 1.1.3.4 0 .7-.1 1-.2.7-.4 1.1-1.2.9-2-.2-.6-.5-.9-1.6-1.7-1-.7-1.3-.9-1.7-1.3-.7-.7-.8-1.6-.4-2.4.2-.5.5-.8 1-1 .4-.2.7-.3 1.3-.3.6 0 1.1.1 1.4.2.1 0 .2.1.3.1 0 0 .1-.1.2-.5l.1-.5h.2s0 .2-.1.5c-.1.8-.2 1.7-.2 2.3 0 .5 0 .5-.1.5-.2-.1-.2 0-.2-.5 0-.6-.1-.8-.4-1.1-.3-.3-.5-.4-.9-.5-.3-.1-.9 0-1.1 0-.7.2-1.1.7-1 1.4 0 .6.3.9 1.8 1.9.8.5 1.1.8 1.4 1.1.4.4.6.8.7 1.3.1.3 0 1-.1 1.2-.3 1-1.1 1.7-2.3 1.9-.6.1-1.5 0-2.2-.2-.1 0-.3-.1-.3-.1s-.1.3-.1.6c-.2.7-.1.7-.2.7 0 .2 0 .2 0 0z" />
-  </svg>
-);
-
-export default PixarIcon;
