--- src/icons/SquaresIcon.js
@@ -1,11 +1,0 @@
-const RocketIcon = ({ className }) => (
-  <svg
-    viewBox="-19 -19 600 600"
-    xmlns="http://www.w3.org/2000/svg"
-    className={className}
-  >
-    <path d="M251 13c0-7-5-13-12-13H13C6 0 0 6 0 13v226c0 7 6 12 13 12h226c7 0 12-5 12-12zm-25 213H25V25h201zm0 0M563 13c0-7-6-13-13-13H324c-7 0-13 6-13 13v226c0 7 6 12 13 12h226c7 0 13-5 13-12zm-25 213H336V25h202zm0 0M251 324c0-7-5-13-12-13H13c-7 0-13 6-13 13v226c0 7 6 13 13 13h226c7 0 12-6 12-13zm-25 212H25V336h201zm0 0M563 324c0-7-6-13-13-13H324c-7 0-13 6-13 13v226c0 7 6 13 13 13h226c7 0 13-6 13-13zm-25 212H336V336h202zm0 0" />
-  </svg>
-);
-
-export default RocketIcon;
