--- src/icons/TargetIcon.js
@@ -1,12 +1,0 @@
-const TargetIcon = ({ className }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 286.5 63"
-    className={className}
-  >
-    <path d="M31.51 42.141c5.774 0 10.455-4.683 10.455-10.455 0-5.773-4.681-10.455-10.455-10.455-5.771 0-10.453 4.682-10.453 10.455 0 5.772 4.682 10.455 10.453 10.455zm0 9.872c11.119 0 20.132-9.015 20.132-20.134 0-11.118-9.013-20.13-20.132-20.13-11.117 0-20.131 9.012-20.131 20.13-.001 11.119 9.014 20.134 20.131 20.134zM1.198 31.879c0-16.74 13.571-30.31 30.312-30.31 16.74 0 30.313 13.57 30.313 30.31 0 16.742-13.573 30.312-30.313 30.312-16.741-.001-30.312-13.57-30.312-30.312z" />
-    <path d="M71.667 10.125h34.539v8.714h-11.88v34.926H83.771v-34.86H71.667v-8.78zM158.224 18.905h-9.683v8.713s7.554.026 8.684 0c3.034-.063 5.38-1.418 5.38-4.39-.001-3.231-1.992-4.323-4.381-4.323zm-19.755-8.78h20.498c11.653 0 13.266 7.973 13.266 12.62 0 6.747-5.81 10.137-6.326 10.266 0 0 5.102 1.904 5.102 7.683v8.134c0 3.262 2.001 4.164 2.354 4.938h-8.327c-3.971-.45-3.905-3.613-3.874-5.422v-7.003c-.097-1.71-.613-3.618-4.196-3.681h-8.424v16.106h-10.103l.03-43.641zM216.78 10.125v43.64h31.438v-8.972h-21.174v-9.751h18.076v-8.261h-18.076v-7.876h21.174v-8.78H216.78zM249.705 18.905h12.461v34.86h10.326v-34.86h12.653v-8.78h-35.44v8.78zM117.115 25.522l-3.712 11.265h7.652l-3.94-11.265zm-5.85-15.405h11.202l15.26 43.648H126.87l-3.152-9.069h-12.881l-2.973 9.069H97.098l14.167-43.648zM213.164 29.396h-18.753v8.488h8.907c-.018 1.747.259 3.49-1.834 5.943-.714.835-3.153 2.708-6.56 2.708-2.872 0-4.937-.821-6.729-2.342-3.959-3.362-4.891-9.642-4.891-12.927 0-4.132 1.645-14.394 11.748-14.815 0 0 6.813-.44 8.589 7.778h9.456c-.192-4.165-5.036-15.862-18.045-15.88-4.49-.006-10.188 1.345-14.906 6.07-3.418 3.422-6.493 9.169-6.493 17.266 0 8.784 3.233 14.494 6.19 17.63 5.19 5.509 10.421 5.933 13.145 5.933 2.161 0 6.295-.321 11.653-3.937l1.419 2.452h7.103V29.396h.001zM277.283 50.896h.543c.347 0 .604-.054.604-.427 0-.333-.352-.384-.633-.384h-.514v.811zm-.611-1.288h1.161c.771 0 1.208.24 1.208.912 0 .565-.354.783-.81.823l.822 1.346h-.638l-.75-1.316h-.383v1.316h-.611v-3.081h.001zm1.096 3.714c1.141 0 2.074-.913 2.074-2.171 0-1.251-.934-2.154-2.074-2.154-1.157 0-2.076.903-2.076 2.154-.001 1.258.918 2.171 2.076 2.171zm0-4.916c1.492 0 2.824 1.075 2.824 2.745 0 1.677-1.332 2.758-2.824 2.758-1.494 0-2.826-1.081-2.826-2.758-.001-1.67 1.331-2.745 2.826-2.745z" />
-  </svg>
-);
-
-export default TargetIcon;
