--- src/icons/TedIcon.js
@@ -1,11 +1,0 @@
-const TedIcon = ({ className }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 512 512"
-    className={className}
-  >
-    <path d="M51.8 210.9H0v-47.4h160.7v47.4h-51.8v137.6h-57l-.1-137.6zm117.7-47.4h155.9v47.4h-98.9V234h98.9v44h-98.9v23.1h98.9v47.4H169.5v-185zM392 301.2h22.3c35.5 0 40.7-28.8 40.7-46.2 0-11.6-3.8-44-45.1-44h-18.3v90.2h.4zm-57-137.7h93.6c61.7 0 83.4 45.6 83.4 92.2 0 56.7-30.1 92.8-94.6 92.8H335v-185z" />
-  </svg>
-);
-
-export default TedIcon;
