--- src/icons/YahooIcon.js
@@ -1,11 +1,0 @@
-const YahooIcon = ({ className }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 512 512"
-    className={className}
-  >
-    <path d="M122.3 209.1c-3.9 1.4-7.1 1.3-10.9 0-12.6 31.6-27.1 67.2-40.1 93.8h.1-.1c6-1.5 11.7-1 15.4 0 3.3-8.4 7.1-17.9 10.9-27.5 12.8-.4 25.6-.4 38.4 0 3.6 9.2 7.3 18.3 10.9 27.5 5.7-1.5 11.4-1.1 15.6 0-9.4-17.6-28.2-63.5-40.2-93.8zm11.2 59.7c-11.1.4-22.2.4-33.2 0 6.8-17.2 13.2-33.8 16.6-43.6 5.1 14.5 10.8 29 16.6 43.6zm-62.2 34.1zm22.6-106.3c-5.7 1.3-10.7 1.3-15.8 0-4.5 8.3-20.8 35-31.3 52.2-10.6-17.5-23.1-37.7-31.3-52.2-6.5 1.4-9.2 1.5-15.6 0 12.8 19.3 33.3 56.1 40.3 68.1l-.9 50.3s4.5-.7 7.5-.7c3.3 0 7.5.7 7.5.7l-.9-50.3c13.1-22.8 34.6-60.1 40.5-68.1zm210.6 11.2c-33.2 0-50.5 21-50.5 48.3 0 29.8 19 48 50.5 48 34.1 0 50.7-20.5 50.7-47.7-.1-30.3-19.9-48.6-50.7-48.6zm.1 91c-22.1 0-35.1-15.8-35.1-42.6 0-30.2 16-42.4 34.2-42.7h.9c20.3 0 35.2 13.8 35.2 42.7-.1 28.8-15 42.6-35.2 42.6zm182.9 1.6c-4.9 0-7.4 3.7-7.4 7.8 0 4.8 3.8 7.2 8.5 7.2 3.4 0 7.1-2.1 7.1-7.7 0-4.5-3.5-7.3-8.2-7.3zm18-102.4l-1.6-.1c-3.7 0-7.5 2.2-7.5 7.8-2.2 28.2-4.5 56.5-8.4 83.4 2.3-.4 4.1.2 5.8.9 4.3-27.3 14.6-68.3 17.2-79.9.4-1.6.9-3.6.9-4.9.1-4.2-2.5-6.6-6.4-7.2zm-86.6 4.2c-33 0-58.2 18.8-58.2 53.6 0 28.7 16.8 54 58.5 54 35.3 0 58.4-19.4 58.4-53.5 0-32.1-21.5-54.1-58.7-54.1zm.1 101.9c-22.4 0-42.7-13.1-42.7-48.5 0-36.7 22.6-47.6 42.7-47.6 24.8 0 43.1 15.1 43.1 48.3-.1 36.1-22.2 47.8-43.1 47.8zm-185.2-95c.8 14.6 1.1 29.2 1.2 43.8-17.5.4-34.9.4-52.4 0 0-14.6.4-29.2 1.2-43.8-5.1 1.2-10.3 1.3-15.6 0 1.4 31.3 1.2 62.6 0 93.8 5.1-1.2 10.3-1.3 15.6 0-.8-14.6-1.1-29.1-1.2-43.6 17.4-.2 34.9-.3 52.4 0 0 14.5-.4 29.1-1.2 43.6 5.3-1.3 10.5-1.2 15.6 0-1.2-31.2-1.4-62.4 0-93.8-5.3 1.3-10.5 1.2-15.6 0z" />
-  </svg>
-);
-
-export default YahooIcon;
